models/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
coloredlogs==15.0.1
flatbuffers==25.2.10
humanfriendly==10.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==1.24.3
onnx==1.14.0
onnx-simplifier==0.4.28
onnxruntime==1.14.1
packaging==25.0
pillow==11.3.0
protobuf==6.31.1
Pygments==2.19.2
requests==2.32.4
rich==14.0.0
sympy==1.14.0
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
typing_extensions==4.14.1
urllib3==2.5.0
